requirements.txt
numpy
pandas
scipy
matplotlib
torch
torchvision
Pillow
scikit-learn
